+++ brewtils/rest/system_client.py
@@ -423,7 +423,10 @@ class SystemClient(object):
 
             if not self.target_self:
                 request = self._easy_client.create_request(
-                    request, blocking=blocking, timeout=timeout
+                    request,
+                    blocking=blocking,
+                    timeout=timeout,
+                    target_garden=self._system.garden_name,
                 )
 
         except ValidationError: